Added unit test for JsonConfigLoader
